Meghna-Query_Classifier/requirements.txt
mistralai
python-dotenv
google-generativeai
azure-ai-inference
azure-core